requirements.txt
crewai
langchain-google-genai
chromadb
llama-parse
python-dotenv
langchain
langchain-community
nest_asyncio
duckduckgo-search
